+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     mark the node as leaf node.
   * [Added] ext-filter methods node.isMatched() and tree.isFilterActive()
   * [Added] CSS for ext-childcounter badges is now part of the themes
+  * [Added] ext-childcounter method `node.updateCounter()`
 
 # 2.12.0 / 2015-09-10
   * [Changed] Documented `iconClass` callback and changed signature from